requirements.txt
PyMuPDF>=1.23.0
faiss-cpu>=1.7.4
sentence-transformers>=2.5.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
streamlit>=1.35.0
numpy>=1.24.0
requests>=2.31.0